Add unit test for server and client.

+++ BoostAsioClient/Include/BoostAsioClient/SyncClient.h
@@ -38,7 +38,7 @@ namespace BoostAsioClient {
         tcp::socket socket_;
         enum { kBuffSize = 1024 };
         char buff_[kBuffSize];
-        int readedSize_;
+        size_t readedSize_;
         bool started_;
         std::string userName_;
     };
